Reorganize and rename.
